+++ src/App.vue
@@ -3,7 +3,10 @@
     id="app"
     @keyup="keyup"
     tabindex="-1"
-    :class="{ night: grimoire.isNight }"
+    :class="{
+      night: grimoire.isNight,
+      static: grimoire.isStatic
+    }"
     :style="{
       backgroundImage: grimoire.background
         ? `url('${grimoire.background}')`
@@ -202,6 +205,14 @@ ul {
   align-items: center;
   align-content: center;
   justify-content: center;
+
+  // disable all animations
+  &.static *,
+  &.static *:after,
+  &.static *:before {
+    transition: none !important;
+    animation: none !important;
+  }
 }
 
 #version {